moved the four classes into RUFAS/ and deleted RUFAS/classes/
