+++ Example-animal_management_SA-wrapper.py
@@ -3,10 +3,10 @@ import json
 from pathlib import Path
 from typing import Dict, Tuple, Any
 
-from RUFAS.classes.config import Config
+from RUFAS.config import Config
 from RUFAS.routines.feed import Feed
-from RUFAS.classes.weather import Weather
-from RUFAS.classes.time import Time
+from RUFAS.weather import Weather
+from RUFAS.time import Time
 from RUFAS.routines.animal.animal_manager import AnimalManager
 
 """
