+++ RUFAS/simulation_engine.py
@@ -4,10 +4,10 @@ import sys
 import time as timer
 
 from RUFAS import routines
-from RUFAS.classes.config import Config
-from RUFAS.classes.state import State
-from RUFAS.classes.weather import Weather
-from RUFAS.classes.time import Time
+from RUFAS.config import Config
+from RUFAS.state import State
+from RUFAS.weather import Weather
+from RUFAS.time import Time
 from RUFAS.output_manager import OutputManager
 from RUFAS.input_manager import InputManager
 from RUFAS.routines.manure.manure_manager import simulate_daily_manure_manager
