+++ tests/soil_crop_tests/manager_tests/test_field_manager.py
@@ -9,8 +9,8 @@ from RUFAS.routines.field.field.field import Field
 from RUFAS.routines.field.soil.layer_data import LayerData
 from RUFAS.routines.field.soil.soil import Soil
 from RUFAS.routines.field.soil.soil_data import SoilData
-from RUFAS.classes.time import Time
-from RUFAS.classes.weather import Weather
+from RUFAS.time import Time
+from RUFAS.weather import Weather
 from RUFAS.routines.manure.manure_manager import ManureManager
 import pytest
 from typing import List, Dict
