+++ tests/test_weather.py
@@ -1,6 +1,6 @@
 import pytest
 
-from RUFAS.classes.weather import Weather
+from RUFAS.weather import Weather
 
 
 @pytest.mark.parametrize("avg_daily_temperatures,expected", [
